only show course children in sorted order
This should be the starting point for a complete table of contents in collapsible form. With this step only the first level items (same as the items on the first page of the box-menu) are shown. They are in the same sort order as in the course. If via authoring tool the sort order of these menu items is changed, the sort order in the drawer will change accordingly.

In a further step the level 2 items (children of the items shown by this commit) shall be implemented.